fixup! refactor(oauth-provider): merge session and device concepts
