+++ packages/oauth-provider/src/device/device-manager.ts
