+++ packages/oauth-provider/src/oauth-provider.ts
@@ -46,8 +46,8 @@ import { ClientStore, asClientStore } from './client/client-store.js'
 import { AuthEndpoint, Client } from './client/client.js'
 import { AUTH_MAX_AGE, TOKEN_MAX_AGE } from './constants.js'
 import { DeviceId } from './device/device-id.js'
+import { DeviceManager } from './device/device-manager.js'
 import { DeviceStore, asDeviceStore } from './device/device-store.js'
-import { DeviceManager } from './device/session-manager.js'
 import { AccessDeniedError } from './errors/access-denied-error.js'
 import { AccountSelectionRequiredError } from './errors/account-selection-required-error.js'
 import { ConsentRequiredError } from './errors/consent-required-error.js'
